moved utils/config.go to config/config.go
