+++ internal/tasks/tasks_test.go
@@ -16,6 +16,7 @@ import (
 
 	"purch/internal/database"
 	"purch/internal/utils"
+	"purch/internal/config"
 )
 
 const (
@@ -124,7 +125,7 @@ func TestMain(m *testing.M) {
 
 func TestSyncItem(t *testing.T) {
 	plaidClient := utils.GetPlaidClient()
-	config := utils.GetConfig()
+	config := config.GetConfig()
 	ctx := context.Background()
 	// create sandbox item
 	publicTokenExchangeResp := createSandboxItem(t, ctx, plaidClient, FIRST_PLATYPUS_BANK, config.GetPlaidProducts())
@@ -155,7 +156,7 @@ func TestSyncItem(t *testing.T) {
 func TestSyncAccounts(t *testing.T) {
 	// setup
 	plaidClient := utils.GetPlaidClient()
-	config := utils.GetConfig()
+	config := config.GetConfig()
 	ctx := context.Background()
 	// create sandbox item
 	publicTokenExchangeResp := createSandboxItem(t, ctx, plaidClient, FIRST_PLATYPUS_BANK, config.GetPlaidProducts())
@@ -186,7 +187,7 @@ func TestSyncAccounts(t *testing.T) {
 func TestSyncTransactions(t *testing.T) {
 	// setup
 	plaidClient := utils.GetPlaidClient()
-	config := utils.GetConfig()
+	config := config.GetConfig()
 	ctx := context.Background()
 	// create sandbox item
 	publicTokenExchangeResp := createSandboxItem(t, ctx, plaidClient, FIRST_PLATYPUS_BANK, config.GetPlaidProducts())
